Isolate CPython page SSR chunk and slim HTML-to-markdown deps
